优化部分界面显示，优化部分代码逻辑，解决了搜索页点击图片无法跳转的问题，增强了代码的异常处理能力，添加了更多人性化提示信息

+++ BrilliantComic/ViewModels/HistoryViewModel.cs
@@ -1,11 +1,13 @@
 ﻿using BrilliantComic.Models.Comics;
 using BrilliantComic.Models.Enums;
 using BrilliantComic.Services;
+using CommunityToolkit.Maui.Alerts;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
+using System.Diagnostics;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -53,14 +55,19 @@ namespace BrilliantComic.ViewModels
         /// 清空历史漫画
         /// </summary>
         /// <returns></returns>
-        [RelayCommand]
-        private async Task CleanComicsAsync()
+        public async Task ClearHistoryComicsAsync()
         {
+            if(Comics.Count == 0)
+            {
+                _ = Toast.Make("暂无历史记录").Show();
+                return;
+            }
             foreach (var item in Comics)
             {
                 await _db.DeleteComicAsync(item, item.Category);
             }
             Comics.Clear();
+            _ = Toast.Make("历史记录已清空").Show();
         }
 
         /// <summary>
